requirements.txt
torch>=1.10.0
torchvision>=0.11.0
pillow>=9.0.0
